+++ docs/changelog.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.5.5] - 2021-06-28
+
+  - Fix steinbock version check
+  - Add meanfactor option to Ilastik (compatibility with IMC segmentation pipeline)
+  - Save Ilastik images/crops as uint16 (compatibility with IMC segmentation pipeline)
+
+
 ## [0.5.4] - 2021-06-23
 
   - Speed up Ilastik image creation
@@ -128,6 +135,7 @@ Added:
 Initial release for beta testing
 
 
+[0.5.5]: https://github.com/BodenmillerGroup/steinbock/compare/v0.5.4...v0.5.5
 [0.5.4]: https://github.com/BodenmillerGroup/steinbock/compare/v0.5.3...v0.5.4
 [0.5.3]: https://github.com/BodenmillerGroup/steinbock/compare/v0.5.2...v0.5.3
 [0.5.2]: https://github.com/BodenmillerGroup/steinbock/compare/v0.5.1...v0.5.2